refactor: split god-object album-management and auth-extended route files (ticket #1506)
Break the two largest route files into cohesive sub-routers plus a shared
module, mounted on the same base routers so /api/albums and /api/auth-extended
URLs are unchanged. Handler bodies moved verbatim; no behaviour changes.

album-management.ts -> crud / photos / optimization / organization / video
  sub-routers + album-management-shared.ts.
auth-extended.ts -> login / invites / password-reset / mfa / passkeys / users
  sub-routers + auth-extended-shared.ts (rate limiter + challenge Map singleton).
