+++ backend/src/routes/album-management-optimization.ts
@@ -1,0 +1,109 @@
+/**
+ * Album Management — optimization queueing sub-router.
+ * Extracted from album-management.ts (ticket #1506).
+ */
+import { Router, Request, Response, NextFunction } from "express";
+import fs from "fs";
+import path from "path";
+import { fileURLToPath } from "url";
+import { execFile, spawn } from "child_process";
+import { promisify } from "util";
+import multer from "multer";
+import os from "os";
+import sharp from "sharp";
+import { csrfProtection } from "../security.js";
+import { requireAuth, requireAdmin, requireManager } from '../auth/middleware.js';
+import { sendNotificationToUser } from '../push-notifications.js';
+import { translateNotification } from '../i18n-backend.js';
+import { getAllUsers } from '../database-users.js';
+import { 
+  deleteAlbumMetadata, 
+  deleteImageMetadata, 
+  saveAlbum, 
+  deleteAlbumState,
+  setAlbumPublished,
+  setAlbumShowOnHomepage,
+  setAlbumDescription,
+  updateImageSortOrder,
+  saveImageMetadata,
+  updateAlbumSortOrder,
+  getAlbumState,
+  getDatabase,
+  setAlbumFolder,
+  getAlbumsInFolder,
+  setFolderPublished,
+  renameAlbum
+} from "../database.js";
+import { processVideo, VideoProcessingProgress } from "../utils/video-processor.js";
+import { invalidateAlbumCache } from "./albums.js";
+import { generateStaticJSONFiles } from "./static-json.js";
+import { generateHomepageHTML } from "./homepage-html.js";
+import { broadcastOptimizationUpdate, queueOptimizationJob } from "./optimization-stream.js";
+import OpenAI from "openai";
+import { error, warn, info, debug, verbose } from '../utils/logger.js';
+import {
+  notifyAllAdmins,
+  trackPhotoUpload,
+  toTitleCase,
+  generateAITitleForImageAsync,
+  generateAITitleForImage,
+  upload,
+  cleanupUploadedTempFile,
+  sanitizeName,
+  sanitizePhotoName,
+  isVideoFile,
+  isPathWithinDirectory,
+  execFileAsync,
+  __dirname,
+  __filename,
+} from "./album-management-shared.js";
+
+const router = Router();
+
+router.post("/:album/optimize", requireAdmin, async (req: Request, res: Response): Promise<void> => {
+  try {
+    const { album } = req.params;
+    
+    const sanitizedAlbum = sanitizeName(album);
+    if (!sanitizedAlbum) {
+      res.status(400).json({ error: 'Invalid album name' });
+      return;
+    }
+
+    // Get project root (two levels up from backend/src)
+    const projectRoot = path.resolve(__dirname, '../../../');
+    const scriptPath = path.join(projectRoot, 'scripts', 'optimize_all_images.js');
+
+    if (!fs.existsSync(scriptPath)) {
+      res.status(500).json({ error: 'Optimization script not found' });
+      return;
+    }
+
+    // Run optimization script in the background using execFile to prevent command injection
+    // Don't wait for it to complete
+    const photosDir = req.app.get("photosDir");
+    const albumPath = path.join(photosDir, sanitizedAlbum);
+    
+    execFile('node', [scriptPath, albumPath], (err, stdout, stderr) => {
+      if (err) {
+        error('[AlbumManagement] Optimization error:', err);
+      } else {
+        info('[AlbumManagement] Optimization complete for album:', sanitizedAlbum);
+      }
+    });
+
+    res.json({ 
+      success: true, 
+      message: 'Optimization started in background' 
+    });
+  } catch (err) {
+    error('[AlbumManagement] Failed to trigger optimization:', err);
+    res.status(500).json({ error: 'Failed to trigger optimization' });
+  }
+});
+
+/**
+ * Update photo order in an album
+ */
+
+export default router;
